one more int64?

+++ tests/diffraction/filtering_test.py
@@ -33,7 +33,9 @@ def make_data_with_pulse_time(rng, n_event):
                           })
     return sc.binning.make_binned(
         events,
-        edges=[sc.array(dims=['tof'], values=[10, 500, 1000], unit='us')],
+        edges=[
+            sc.array(dims=['tof'], values=[10, 500, 1000], unit='us', dtype='int64')
+        ],
         groups=[sc.arange('spectrum', 0, 10, unit=None, dtype='int64')])
